moved ImageBaseMatcher class to matcher.py file
now the base class and the actual matcher implementations are in the same file
